fix : ReservationResponseDto > amount 에 대해 long 에서 bigdecimal 로 변경

+++ src/main/java/com/example/cinemareservation/domain/cinema/dto/response/ReservationResponseDto.java
@@ -1,12 +1,14 @@
 package com.example.cinemareservation.domain.cinema.dto.response;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 public record ReservationResponseDto(
-        String movieName,
-        Timestamp timestamp,
-        Integer personCount,
-        Long originalPaymentAmount,
-        Long totalPaymentAmount
+	String movieName,
+	Timestamp timestamp,
+	Integer personCount,
+	BigDecimal originalPaymentAmount,
+	BigDecimal totalPaymentAmount
 ) {
+
 }
